Se separan responsabilidades en modulos separados.
