+++ UAO-Neumonia/grad_cam.py
@@ -1,0 +1,36 @@
+# Módulo encargado de generar Grad-CAM y superponerlo en la imagen original
+import numpy as np
+import cv2
+import tensorflow as tf
+import tensorflow.keras.backend as K
+
+from load_model import model_fun
+from preprocess_img import preprocess
+
+def grad_cam(array):
+    img = preprocess(array)
+    model = model_fun()
+    preds = model.predict(img)
+    argmax = np.argmax(preds[0])
+    output = model.output[:, argmax]
+    last_conv_layer = model.get_layer("conv10_thisone")
+    grads = K.gradients(output, last_conv_layer.output)[0]
+    pooled_grads = K.mean(grads, axis=(0, 1, 2))
+    iterate = K.function([model.input], [pooled_grads, last_conv_layer.output[0]])
+    pooled_grads_value, conv_layer_output_value = iterate(img)
+    for filters in range(64):
+        conv_layer_output_value[:, :, filters] *= pooled_grads_value[filters]
+    # creating the heatmap
+    heatmap = np.mean(conv_layer_output_value, axis=-1)
+    heatmap = np.maximum(heatmap, 0)  # ReLU
+    heatmap /= np.max(heatmap)  # normalize
+    heatmap = cv2.resize(heatmap, (img.shape[1], img.shape[2]))
+    heatmap = np.uint8(255 * heatmap)
+    heatmap = cv2.applyColorMap(heatmap, cv2.COLORMAP_JET)
+    img2 = cv2.resize(array, (512, 512))
+    hif = 0.8
+    transparency = heatmap * hif
+    transparency = transparency.astype(np.uint8)
+    superimposed_img = cv2.add(transparency, img2)
+    superimposed_img = superimposed_img.astype(np.uint8)
+    return superimposed_img[:, :, ::-1]
