+++ UAO-Neumonia/integrator.py
@@ -1,0 +1,24 @@
+# Módulo integrador que se encarga de orquestar las funciones de lectura de imagen,
+import os
+import cv2
+from read_img import read_dicom_file
+from load_model import get_model
+
+def predict(array):
+    #   1. call function to pre-process image: it returns image in batch format
+    batch_array_img = preprocess(array)
+    #   2. call function to load model and predict: it returns predicted class and probability
+    model = model_fun()
+    # model_cnn = tf.keras.models.load_model('conv_MLP_84.h5')
+    prediction = np.argmax(model.predict(batch_array_img))
+    proba = np.max(model.predict(batch_array_img)) * 100
+    label = ""
+    if prediction == 0:
+        label = "bacteriana"
+    if prediction == 1:
+        label = "normal"
+    if prediction == 2:
+        label = "viral"
+    #   3. call function to generate Grad-CAM: it returns an image with a superimposed heatmap
+    heatmap = grad_cam(array)
+    return (label, proba, heatmap)
