+++ UAO-Neumonia/main.py
@@ -1,122 +1,19 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-#estado inicial
-# esta es la rama main
-
+# script principal de la interfaz gráfica y de la integración de las funciones.
 
 import os
-from tkinter import *
-from tkinter import ttk, font, filedialog, Entry
-
-from tkinter.messagebox import askokcancel, showinfo, WARNING
-import getpass
-import tensorflow as tf
-from PIL import ImageTk, Image
 import csv
-import pyautogui
+import tkinter as tk
+from tkinter import END, Image, StringVar, Text, Tk, ttk, font, filedialog
+from PIL import ImageTk, Image
+from tkinter.messagebox import askokcancel, showinfo, WARNING
+
+from PIL import ImageTk
 import tkcap
-import img2pdf
-import numpy as np
-import time
-tf.compat.v1.disable_eager_execution()
-tf.compat.v1.experimental.output_all_intermediates(True)
-import cv2
-import pydicom as dicom
 
-import tensorflow.keras.backend as K
-from tensorflow.keras.models import load_model
+from detector_neumonia import mostrarDato, predict, read_dicom_file, read_jpg_file
+import load_model
 
-
-# fUNCION PARA MOSTRAR UNA VENTANA DE DIALOGO
-def mostrarDato(filepath):
-    root = Tk()
-    root.withdraw()  # Ocultar la ventana principal
-
-    # Mostrar el cuadro de mensaje
-    showinfo("Información", filepath)
-
-    # Cerrar la ventana principal
-    root.destroy()
-
-def grad_cam(array):
-    img = preprocess(array)
-    model = model_fun()
-    preds = model.predict(img)
-    argmax = np.argmax(preds[0])
-    output = model.output[:, argmax]
-    last_conv_layer = model.get_layer("conv10_thisone")
-    grads = K.gradients(output, last_conv_layer.output)[0]
-    pooled_grads = K.mean(grads, axis=(0, 1, 2))
-    iterate = K.function([model.input], [pooled_grads, last_conv_layer.output[0]])
-    pooled_grads_value, conv_layer_output_value = iterate(img)
-    for filters in range(64):
-        conv_layer_output_value[:, :, filters] *= pooled_grads_value[filters]
-    # creating the heatmap
-    heatmap = np.mean(conv_layer_output_value, axis=-1)
-    heatmap = np.maximum(heatmap, 0)  # ReLU
-    heatmap /= np.max(heatmap)  # normalize
-    heatmap = cv2.resize(heatmap, (img.shape[1], img.shape[2]))
-    heatmap = np.uint8(255 * heatmap)
-    heatmap = cv2.applyColorMap(heatmap, cv2.COLORMAP_JET)
-    img2 = cv2.resize(array, (512, 512))
-    hif = 0.8
-    transparency = heatmap * hif
-    transparency = transparency.astype(np.uint8)
-    superimposed_img = cv2.add(transparency, img2)
-    superimposed_img = superimposed_img.astype(np.uint8)
-    return superimposed_img[:, :, ::-1]
-
-
-def predict(array):
-    #   1. call function to pre-process image: it returns image in batch format
-    batch_array_img = preprocess(array)
-    #   2. call function to load model and predict: it returns predicted class and probability
-    model = model_fun()
-    # model_cnn = tf.keras.models.load_model('conv_MLP_84.h5')
-    prediction = np.argmax(model.predict(batch_array_img))
-    proba = np.max(model.predict(batch_array_img)) * 100
-    label = ""
-    if prediction == 0:
-        label = "bacteriana"
-    if prediction == 1:
-        label = "normal"
-    if prediction == 2:
-        label = "viral"
-    #   3. call function to generate Grad-CAM: it returns an image with a superimposed heatmap
-    heatmap = grad_cam(array)
-    return (label, proba, heatmap)
-
-
-def read_dicom_file(path):
-    img = dicom.read_file(path)
-    img_array = img.pixel_array
-    img2show = Image.fromarray(img_array)
-    img2 = img_array.astype(float)
-    img2 = (np.maximum(img2, 0) / img2.max()) * 255.0
-    img2 = np.uint8(img2)
-    img_RGB = cv2.cvtColor(img2, cv2.COLOR_GRAY2RGB)
-    return img_RGB, img2show
-
-
-def read_jpg_file(path):
-    img = cv2.imread(path)
-    img_array = np.asarray(img)
-    img2show = Image.fromarray(img_array)
-    img2 = img_array.astype(float)
-    img2 = (np.maximum(img2, 0) / img2.max()) * 255.0
-    img2 = np.uint8(img2)
-    return img2, img2show
-
-
-def preprocess(array):
-    array = cv2.resize(array, (512, 512))
-    array = cv2.cvtColor(array, cv2.COLOR_BGR2GRAY)
-    clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(4, 4))
-    array = clahe.apply(array)
-    array = array / 255
-    array = np.expand_dims(array, axis=-1)
-    array = np.expand_dims(array, axis=0)
-    return array
+# Importamos la función de integración:
 
 def model_fun():
     # Carga un modelo ya entrenado
@@ -130,7 +27,7 @@ def model_fun():
 class App:
     def __init__(self):
         self.root = Tk()
-        self.root.title("Herramienta para la detección rápida de neumonía")
+        self.root.title("(MAIN)Herramienta para la detección rápida de neumonía")
 
         #   BOLD FONT
         fonti = font.Font(weight="bold")
